@@ -36,33 +36,34 @@
     },
     "homepage": "https://iherman.github.io/r2epub/",
     "dependencies": {
+        "@exodus/schemasafe": "^1.3.0",
         "ajv": "^6.12.6",
-        "commander": "^5.1.0",
-        "jsdom": "^19.0.0",
-        "jszip": "^3.2.2",
-        "minimist": "^1.2.5",
+        "commander": "^14.0.0",
+        "jsdom": "^26.1.0",
+        "jszip": "^3.10.1",
+        "minimist": "^1.2.8",
         "node-fetch": "^2.6.1",
         "underscore": "^1.10.2",
-        "url": "^0.11.0",
+        "url": "^0.11.4",
         "valid-url": "^1.0.9",
-        "w3c-xmlserializer": "^3.0.0",
-        "xmlbuilder2": "^1.4.3"
+        "w3c-xmlserializer": "^5.0.0",
+        "xmlbuilder2": "^3.1.1"
     },
     "devDependencies": {
         "@types/ajv": "^1.0.0",
-        "@types/commander": "^2.12.2",
-        "@types/jsdom": "^16.2.14",
-        "@types/jszip": "^3.1.7",
-        "@types/node": "^13.7.7",
+        "@types/commander": "^2.12.0",
+        "@types/jsdom": "^21.1.7",
+        "@types/jszip": "^3.4.0",
+        "@types/node": "^22.15.21",
         "@types/node-fetch": "^2.5.5",
         "@types/underscore": "^1.9.4",
-        "@types/valid-url": "^1.0.2",
-        "@types/xmlserializer": "^0.6.0",
+        "@types/valid-url": "^1.0.7",
+        "@types/w3c-xmlserializer": "^2.0.4",
         "@typescript-eslint/eslint-plugin": "^4.6.1",
         "@typescript-eslint/parser": "^4.6.1",
         "eslint": "^7.13.0",
         "terser": "^4.6.11",
         "typedoc": "^0.15.8",
-        "typescript": "^4.5.4"
+        "typescript": "^4.9.5"
     }
 }
